requirements.txt
numpy>=1.20.0
torch>=1.9.0
gymnasium>=0.26.0
pybullet>=3.2.1
matplotlib>=3.4.0
pyyaml>=6.0
pandas>=1.3.0
seaborn>=0.11.0
tensorboard>=2.6.0
tqdm>=4.62.0
pytest>=6.2.5
scikit-learn>=0.24.2 